requirements.txt
absl-py==2.1.0
accelerate==1.4.0
autocommand==2.2.2
av==14.2.0
backports.tarfile==1.2.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.1.0a1
decord==0.6.0
flatbuffers==25.2.10
fonttools==4.56.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mkl-service==2.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.0.0
platformdirs==4.2.2
protobuf==6.30.0
psutil==7.0.0
pybind11==2.13.6
pycocotools==2.0.8
pyparsing==3.2.1
regex==2024.11.6
safetensors==0.5.3
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
transformers==4.49.0
triton==3.2.0
typeguard==4.3.0
zipp==3.19.2
